Move recent patient view outside patient package
